backend/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-groq
pydantic
python-dotenv
langchain-community
tavily-python
langgraph